- Merged comprehensive algorithmic optimizations
- Integrated advanced Linux/BSD compatibility layers
- Added enhanced security, networking, and driver frameworks
- Resolved all conflicts in favor of optimized implementations

+++ src/network/tcp_udp.rs
@@ -213,7 +213,7 @@ impl TCPConnection for SimpleSocket {
     }
 
     fn get_state(&self) -> TCPState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) }
     }
 }
